+++ src/test/java/racinggame/domain/AlwaysMoveCondition.java
@@ -1,4 +1,4 @@
-package racinggame;
+package racinggame.domain;
 
 import racinggame.domain.MoveCondition;
