+++ src/test/java/racinggame/domain/CarTest.java
@@ -1,4 +1,4 @@
-package racinggame;
+package racinggame.domain;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
